+++ src/core/event.h
@@ -5,17 +5,17 @@ namespace swarmnet_sim {
 
 class Event {
    public:
-    float get_exec_time() const;
+    double get_exec_time() const;
     int get_from_id() const;
     int get_to_id() const;
     void log_node(int id);
     virtual void exec();
-    Event(void* arena, float exec_time, int from_id, int to_id);
+    Event(void* arena, double exec_time, int from_id, int to_id);
     Event();
 
    protected:
     void* arena;
-    float exec_time;
+    double exec_time;
     int from_id;
     int to_id;
 };
