+++ src/core/node.cpp
@@ -21,15 +21,15 @@ color_t Node::get_color() const { return this->color; }
 
 int Node::get_node_id() const { return this->node_id; }
 
-float Node::get_velocity() const { return this->velocity; }
+double Node::get_velocity() const { return this->velocity; }
 
-float Node::get_radius() const { return this->radius; }
+double Node::get_radius() const { return this->radius; }
 
 bool Node::get_collision_flag() const { return this->collision_flag; }
 
 bool Node::get_skip_logging_flag() const { return this->skip_logging_flag; }
 
-void Node::set_velocity(float velocity) { this->velocity = velocity; }
+void Node::set_velocity(double velocity) { this->velocity = velocity; }
 
 void Node::set_position(position2d_t pos) { this->pos = pos; }
 
@@ -43,7 +43,7 @@ void Node::set_skip_logging_flag(bool skip_logging_flag) {
     this->skip_logging_flag = skip_logging_flag;
 }
 
-void Node::change_theta(float theta_delta) {
+void Node::change_theta(double theta_delta) {
     position2d_t new_pos;
     new_pos.x = pos.x;
     new_pos.y = pos.y;
