+++ src/core/physics_2d.h
@@ -6,9 +6,9 @@
 namespace swarmnet_sim {
 
 typedef struct position {
-    float x;
-    float y;
-    float theta;
+    double x;
+    double y;
+    double theta;
 } position2d_t;
 
 typedef struct color {
@@ -17,15 +17,15 @@ typedef struct color {
     uint8_t blue;
 } color_t;
 
-position2d_t calculate_future_pos(position2d_t start, float v, float seconds);
-float calculate_dist(position2d_t pos1, position2d_t pos2);
+position2d_t calculate_future_pos(position2d_t start, double v, double seconds);
+double calculate_dist(position2d_t pos1, position2d_t pos2);
 
 bool operator==(const position2d_t& lhs, const position2d_t& rhs);
 bool operator==(const color_t& lhs, const color_t& rhs);
 
 class Physics_engine {
    public:
-    virtual float check_collision(float future_time) = 0;
+    virtual double check_collision(double future_time) = 0;
     virtual void init() = 0;
     Physics_engine(void* arena) { this->arena = arena; };
     Physics_engine(){};
