+++ src/plugin/event/kilobot/rx_event.cpp
@@ -12,7 +12,7 @@ void RX_start_event::exec() {
     arena_ptr->get_medium()->start_rx(to_id, this->pkt, this->sensing);
 }
 
-RX_start_event::RX_start_event(void* arena, float exec_time, int to_id,
+RX_start_event::RX_start_event(void* arena, double exec_time, int to_id,
                                packet_t pkt, situated_sensing_t sensing) {
     // std::cout << "RX event created" << std::endl;
     this->arena = arena;
@@ -27,7 +27,7 @@ void RX_end_event::exec() {
     arena_ptr->get_medium()->end_rx(to_id);
 }
 
-RX_end_event::RX_end_event(void* arena, float exec_time, int to_id) {
+RX_end_event::RX_end_event(void* arena, double exec_time, int to_id) {
     this->arena = arena;
     this->exec_time = exec_time;
     this->to_id = to_id;
