+++ src/plugin/medium/kilobot_CSMA_fixed_topology/kilo_fixed_medium.cpp
@@ -12,7 +12,7 @@ namespace swarmnet_sim {
 
 void Kilo_fixed_medium::start_tx(int tx_node_id) {
     // packet_t tx_packet;
-    float comm_radius = COMM_RADIUS;
+    double comm_radius = COMM_RADIUS;
     Arena* arena_ptr = (Arena*)(this->arena);
 
     Node* tx_node = arena_ptr->get_node(tx_node_id);
@@ -31,7 +31,7 @@ void Kilo_fixed_medium::start_tx(int tx_node_id) {
         return;
     }
 
-    float tx_time = sizeof(packet_t) * SPEED_BYTE_PER_SECOND;
+    double tx_time = sizeof(packet_t) * SPEED_BYTE_PER_SECOND;
     // std::cout << "tx time " << tx_time << std::endl;
 
     // std::cout << "start tx for " << tx_node_id << std::endl << std::flush;
@@ -51,7 +51,7 @@ void Kilo_fixed_medium::start_tx(int tx_node_id) {
     for (auto& neighbor : neighbors) {
         Node* rx_node = arena_ptr->get_node(neighbor);
         position2d_t rx_node_pos = rx_node->get_position();
-        float dist = calculate_dist(tx_node_pos, rx_node_pos);
+        double dist = calculate_dist(tx_node_pos, rx_node_pos);
         situated_sensing_t sensing;
         sensing.distance = dist;
         int theta_diff = atan2(tx_node_pos.y - rx_node_pos.y,
@@ -73,11 +73,11 @@ void Kilo_fixed_medium::start_tx(int tx_node_id) {
 void Kilo_fixed_medium::end_tx(int tx_node_id, bool success) {
     // register the next tx
     Arena* arena_ptr = (Arena*)(this->arena);
-    float next_tx_time_noise = ((float)std::rand() / (float)RAND_MAX - 0.5) *
-                               TX_PERIOD_NOISE_RANGE_SECOND;
-    float next_tx_time_local = TX_PERIOD_SECOND + next_tx_time_noise;
+    double next_tx_time_noise = ((double)std::rand() / (double)RAND_MAX - 0.5) *
+                                TX_PERIOD_NOISE_RANGE_SECOND;
+    double next_tx_time_local = TX_PERIOD_SECOND + next_tx_time_noise;
     Node* tx_node = arena_ptr->get_node(tx_node_id);
-    float next_tx_time =
+    double next_tx_time =
         ((Kilobot*)tx_node)->local_time_to_global_time(next_tx_time_local);
     // std::cout << "next_tx_time local " << next_tx_time_local << " global "
     //           << next_tx_time << std::endl;
@@ -107,7 +107,7 @@ void Kilo_fixed_medium::start_rx(int rx_node_id, packet_t rx_packet,
     }
 
     Arena* arena_ptr = (Arena*)(this->arena);
-    float rx_time = sizeof(packet_t) * SPEED_BYTE_PER_SECOND;
+    double rx_time = sizeof(packet_t) * SPEED_BYTE_PER_SECOND;
     // std::cout << "rx time " << rx_time << std::endl;
 
     RX_end_event* rx_end_event = new RX_end_event(
@@ -122,12 +122,12 @@ void Kilo_fixed_medium::end_rx(int rx_node_id) {
     if (this->rx_buffer[rx_node_id].corrupted == false) {
         // not corrupted
         // drop with probability
-        float dice_roll = rand() / float(RAND_MAX);
+        double dice_roll = rand() / double(RAND_MAX);
         int distance = this->rx_buffer[rx_node_id].sensing.distance;
-        float threshold =
-            (float)MAX_SUCCESS_RATE *
-            (1 + (float)DROPPING_SHARPNESS /
-                     (float)(distance - COMM_RADIUS - DROPPING_SHARPNESS));
+        double threshold =
+            (double)MAX_SUCCESS_RATE *
+            (1 + (double)DROPPING_SHARPNESS /
+                     (double)(distance - COMM_RADIUS - DROPPING_SHARPNESS));
         if (dice_roll <= threshold) {
             // no drop
             Node* rx_node = arena_ptr->get_node(rx_node_id);
@@ -145,7 +145,7 @@ void Kilo_fixed_medium::end_rx(int rx_node_id) {
 void Kilo_fixed_medium::init() {
     // compute the fixed topology up front
 
-    float comm_radius = COMM_RADIUS;
+    double comm_radius = COMM_RADIUS;
     Arena* arena_ptr = (Arena*)(this->arena);
     std::cout << "medium init" << std::endl << std::flush;
     const clock_t begin_time = clock();
@@ -159,7 +159,7 @@ void Kilo_fixed_medium::init() {
                 // not self
                 Node* rx_node = arena_ptr->get_node(j);
                 position2d_t rx_node_pos = rx_node->get_position();
-                float dist = calculate_dist(tx_node_pos, rx_node_pos);
+                double dist = calculate_dist(tx_node_pos, rx_node_pos);
                 // std::cout << "dist " << dist << std::endl;
                 if (dist > comm_radius) {
                     // too far
@@ -172,7 +172,7 @@ void Kilo_fixed_medium::init() {
         topology.push_back(neighbors);
     }
     std::cout << "medium init finish "
-              << float(clock() - begin_time) / CLOCKS_PER_SEC << std::endl
+              << double(clock() - begin_time) / CLOCKS_PER_SEC << std::endl
               << std::flush;
 }
